fix(review): robustness guards in sdxl/z-image/flux2/flux (epic 5250 Low tier)
Typed Result errors + shape/bounds guards; no behaviour change on valid input.

sdxl:
- scale_prior_noise: sigmas.last() unwrap -> ok_or_else; GEGLU odd-dim check [sc-5302]
- ip-adapter preprocess: reject zero-dim reference image [sc-5304]

z-image:
- rope_rotate: try_into the 2 outputs; control block0 before_proj ok_or_else [sc-5305]
- token/position shape math: i64 intermediates before usize cast [sc-5306]
- preflight memory guard: conservative caption-token count (was 32) [sc-5307]
- sample_sigma: wrapping_add seed mixing, not XOR-sibling [sc-5308]
- vae encode: reject non-even (2C) channel count [sc-5309]

flux2:
- kv-cache trailing(): guard n>s underflow [sc-5312]
- modulation chunk count / drop_batch B==1 / img_seq underflow guards [sc-5313]

flux:
- vae-remap strip_prefix + image-encoder last() -> typed errors [sc-5314]
- forward_capture: capture block 0 in-loop, drop the recompute [sc-5315]

+++ mlx-gen-z-image/src/pipeline.rs
@@ -210,8 +210,16 @@ pub fn encode_control_context(
     let control_latents = encode_init_latents(vae, control_image, target_width, target_height)?;
     let sh = control_latents.shape(); // [16, 1, H/8, W/8]
     let (c, fdim, h, w) = (sh[0], sh[1], sh[2], sh[3]);
-    let mask = Array::from_slice(&vec![0f32; (fdim * h * w) as usize], &[1, fdim, h, w]);
-    let inpaint = Array::from_slice(&vec![0f32; (c * fdim * h * w) as usize], &[c, fdim, h, w]);
+    // i64 intermediates before the usize cast — bounded by the validated max resolution today, but a
+    // raw i32 product is a latent overflow footgun if the caps change (F-056).
+    let mask = Array::from_slice(
+        &vec![0f32; (fdim as i64 * h as i64 * w as i64) as usize],
+        &[1, fdim, h, w],
+    );
+    let inpaint = Array::from_slice(
+        &vec![0f32; (c as i64 * fdim as i64 * h as i64 * w as i64) as usize],
+        &[c, fdim, h, w],
+    );
     Ok(concatenate_axis(&[&control_latents, &mask, &inpaint], 0)?)
 }
